+++ Spring_SHOOKREAM/src/main/resources/com/itwillbs/shookream/mapper/ProductMapper.xml
@@ -2,103 +2,6 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 <mapper namespace="com.itwillbs.shookream.mapper.ProductMapper">
 	
-	<!-- ============================ 상품 등록 ========================== -->
-	<insert id="insertProduct">
-	<!-- 상품 등록 작업 전 idx 작업 -->
-	<selectKey  keyProperty="product_idx" resultType="int" order="BEFORE">
-		SELECT MAX(product_idx) FROM product
-	</selectKey>
-		INSERT 
-			INTO product
-			VALUES (
-				#{product_idx} + 1
-				,#{product_name}
-				,#{product_brand}
-				,#{product_size}
-				,#{product_price}
-				,#{product_release_price}
-				,#{product_amount}
-				,0
-				,#{product_exp}
-				,#{product_detail_exp}
-				,#{product_color}
-				,#{product_discount_price}
-				,now()
-				,0
-			)
-	</insert>
-	
-	<!-- 상품 이미지 등록 -->
-	<insert id="insertImage">
-	<selectKey keyProperty="image_idx" resultType="int" order="BEFORE">
-	<!-- 상품 이미지 등록 전 idx 작업 -->
-		SELECT MAX(image_idx) FROM image
-	</selectKey>
-		INSERT 
-			INTO image (image_idx, product_idx, image_main_file, image_real_file1)
-			VALUES (
-				(#{image_idx} + 1)
-				,#{product_idx}
-				,#{image.image_main_file}
-				,#{image.image_real_file1}
-<!-- 				,#{image.image_real_file2} -->
-			)
-	</insert>
-	
-	<!-- ============================ 상품 등록 끝 ========================== -->
-	
-	<!-- 상품 수정 -->
-	<update id="modifyProduct">
-		UPDATE product
-  			SET
-  				name = #{product.product_name}
-  				, brand = #{product.product_brand}
-  				, price = #{product.product_price}
-  				, discount = #{product.product_discount_price}
-  				, release_price = #{product.product_release_price}
-  				, size = #{product.product_size}
-  				, amount = #{product.product_amount}
-  				, color = #{product.product_color}
-  				, exp = #{product.product_exp}
-  				, detail_exp = #{product.product_detail_exp}
-  			WHERE 
-  				product_idx = #{product.product_idx}
-	</update>
-	
-	<!-- 상품 삭제 -->
-	<delete id="removeProduct">
-		DELETE
-	  		FROM product
-	  		WHERE product_idx = #{product.product_idx}
-	</delete>
-	
-	<!-- 상품 목록 조회 -->
-
-	<select id="selectProductList" resultType="com.itwillbs.shookream.vo.ProductVo">
-		SELECT 
-			 p.product_idx
-			 , p.product_brand
-			 , p.product_name
-			 , p.product_price
-			 , p.product_date
-			 , p.product_amount
-			 , p.product_color
-			 , i.image_main_file
-			FROM product AS p JOIN image AS i 
-				ON p.product_idx = i.product_idx
-				ORDER BY product_date DESC
-	</select>
-	
-	
-	<!-- 이미지 목록 조회 -->
-<!-- 	<select id="selectImgList" resultType="com.itwillbs.shookream.vo.imageVo"> -->
-<!-- 		SELECT  -->
-<!-- 			image_main_file -->
-<!-- 			FROM image -->
-<!-- 			WHERE product_idx=#{product_idx} -->
-<!-- 	</select> -->
-<!-- ======================================================== -->
-	
 	<select id="selectMemberIdx" resultType="int">
 		SELECT member_idx
 			FROM member
